Rename ItemListView -> BlobListView
The only reason it's using "item" terminology is because I originally
planned on adding this to the v1 API. But even though it's not a
breaking change it's easier to implement this as a V2 resource,
because otherwise I'll need to create two versions, because blobs
are identified using different hashes in v1 and v2.

This does not change the API as /blobs is not in the v1 API at the
moment.
